Event based permissions commit.
- Take advantage of abstract TownyActionEvent class.
- Tidy isAllowedAction().
- Comments.
